move Service from apps/cli/src/services/types.ts into core
