+++ apps/cli/src/core/base-service.ts
@@ -24,7 +24,7 @@
  * and behaviors.
  */
 
-import { Service } from '../services/types.js';
+import { Service } from './service-interface.js';
 import { ServiceName } from './service-discovery.js';
 import { Config, ServiceConfig } from './cli-config.js';
 import { PlatformType } from './platform-types.js';
