+++ apps/cli/src/core/handlers/context.ts
@@ -1,4 +1,4 @@
-import { Service } from '../../services/types.js';
+import { Service } from '../service-interface.js';
 import { BaseHandlerContext } from './types.js';
 
 /**
