+++ apps/cli/src/platforms/container/platform.ts
@@ -22,7 +22,7 @@
 
 import { execSync } from 'child_process';
 import { Platform, LogOptions, LogEntry } from '../../core/platform.js';
-import { Service } from '../../services/types.js';
+import { Service } from '../../core/service-interface.js';
 import { HandlerRegistry } from '../../core/handlers/registry.js';
 import { handlers } from './handlers/index.js';
 import { StateManager } from '../../core/state-manager.js';
