+++ apps/cli/src/platforms/mock/platform.ts
@@ -29,7 +29,7 @@
  */
 
 import { Platform, LogOptions, LogEntry } from '../../core/platform.js';
-import { Service } from '../../services/types.js';
+import { Service } from '../../core/service-interface.js';
 import { HandlerRegistry } from '../../core/handlers/registry.js';
 import { handlers } from './handlers/index.js';
